Uploaded STM32 Project
External Crystall (HSE) is used now. On the project before this, HSI is used. Recheck the Clock Configuration tree. Raw ADC to callibrate, can be sent via USART with the help of the DIP Switch. Succesfully capturing 200 samples, for both AC Voltage and Current. AC Voltage Callibrated and integrated with Type 2 Fuzzy Logic Controller but not yet tested.